+++ plom/baseMessenger.py
@@ -3,12 +3,14 @@
 # Copyright (C) 2019-2021 Colin B. Macdonald
 # Copyright (C) 2021 Peter Lee
 
+from io import BytesIO
+import logging
 import ssl
 import threading
-import logging
 
 import urllib3
 import requests
+from requests_toolbelt import MultipartDecoder
 
 from plom import __version__, Plom_API_Version, Default_Port
 from plom.plom_exceptions import PlomBenignException, PlomSeriousException
@@ -16,6 +18,8 @@ from plom.plom_exceptions import (
     PlomAuthenticationException,
     PlomAPIException,
     PlomExistingLoginException,
+    PlomTaskChangedError,
+    PlomTaskDeletedError,
 )
 
 log = logging.getLogger("messenger")
@@ -489,3 +493,186 @@ class BaseMessenger:
         finally:
             self.SRmutex.release()
         return messenger_response
+
+    def request_ID_images(self, code):
+        self.SRmutex.acquire()
+        try:
+            response = self.session.get(
+                "https://{}/ID/images/{}".format(self.server, code),
+                json={"user": self.user, "token": self.token},
+                verify=False,
+            )
+            response.raise_for_status()
+            imageList = []
+            for img in MultipartDecoder.from_response(response).parts:
+                imageList.append(
+                    BytesIO(img.content).getvalue()
+                )  # pass back image as bytes
+        except requests.HTTPError as e:
+            if response.status_code == 401:
+                raise PlomAuthenticationException() from None
+            elif response.status_code == 404:
+                raise PlomSeriousException(
+                    "Cannot find image file for {}.".format(code)
+                ) from None
+            elif response.status_code == 410:
+                raise PlomBenignException(
+                    "That ID group of {} has not been scanned.".format(code)
+                ) from None
+            elif response.status_code == 409:
+                raise PlomSeriousException(
+                    "Another user has the image for {}. This should not happen".format(
+                        code
+                    )
+                ) from None
+            else:
+                raise PlomSeriousException(
+                    "Some other sort of error {}".format(e)
+                ) from None
+        finally:
+            self.SRmutex.release()
+
+        return imageList
+
+    def request_donotmark_images(self, papernum):
+        """Get the various Do Not Mark images for a paper."""
+        self.SRmutex.acquire()
+        try:
+            response = self.session.get(
+                f"https://{self.server}/ID/donotmark_images/{papernum}",
+                json={"user": self.user, "token": self.token},
+                verify=False,
+            )
+            response.raise_for_status()
+            return [
+                BytesIO(img.content).getvalue()
+                for img in MultipartDecoder.from_response(response).parts
+            ]
+        except requests.HTTPError as e:
+            if response.status_code == 401:
+                raise PlomAuthenticationException() from None
+            elif response.status_code == 404:
+                raise PlomSeriousException(
+                    f"Cannot find DNW image files for {papernum}."
+                ) from None
+            elif response.status_code == 410:
+                raise PlomBenignException(
+                    f"The DNM group of {papernum} has not been scanned."
+                ) from None
+            else:
+                raise PlomSeriousException(
+                    "Some other sort of error {}".format(e)
+                ) from None
+        finally:
+            self.SRmutex.release()
+
+    def get_annotations(self, num, question, edition=None, integrity=None):
+        """Download the latest annotations (or a particular set of annotations).
+
+        Args:
+            num (int): the paper number.
+            question (int): the question number.
+            edition (int/None): which annotation set or None for latest.
+            integrity (str/None): a checksum to ensure the server hasn't
+                changed under us.  Can be omitted if not relevant.
+
+        Returns:
+            dict: contents of the plom file.
+
+        Raises:
+            PlomAuthenticationException
+            PlomTaskChangedError
+            PlomTaskDeletedError
+            PlomSeriousException
+        """
+        if edition is None:
+            edition = "_"  # TODO: use two URLs?
+        if integrity is None:
+            integrity = ""
+        self.SRmutex.acquire()
+        try:
+            response = self.session.get(
+                f"https://{self.server}/MK/annotations/{num}/{question}/{edition}",
+                json={
+                    "user": self.user,
+                    "token": self.token,
+                    "integrity": integrity,
+                },
+                verify=False,
+            )
+            response.raise_for_status()
+            return response.json()
+        except requests.HTTPError as e:
+            if response.status_code == 401:
+                raise PlomAuthenticationException() from None
+            elif response.status_code == 404:
+                raise PlomSeriousException(
+                    "Cannot find image file for {}.".format(num)
+                ) from None
+            elif response.status_code == 406:
+                raise PlomTaskChangedError(
+                    "Task {} has been changed by manager.".format(num)
+                ) from None
+            elif response.status_code == 410:
+                raise PlomTaskDeletedError(
+                    "Task {} has been deleted by manager.".format(num)
+                ) from None
+            else:
+                raise PlomSeriousException(
+                    "Some other sort of error {}".format(e)
+                ) from None
+        finally:
+            self.SRmutex.release()
+
+    def get_annotations_image(self, num, question, edition=None):
+        """Download image of the latest annotations (or a particular set of annotations).
+
+        Args:
+            num (int): the paper number.
+            question (int): the question number.
+            edition (int/None): which annotation set or None for latest.
+
+        Returns:
+            dict: contents of the plom file.
+
+        Raises:
+            PlomAuthenticationException
+            PlomTaskChangedError: TODO: add this back again, with integriy_check??
+            PlomTaskDeletedError
+            PlomSeriousException
+        """
+        if edition is None:
+            edition = "_"  # TODO: use two URLs?
+        self.SRmutex.acquire()
+        try:
+            response = self.session.get(
+                f"https://{self.server}/MK/annotations_image/{num}/{question}/{edition}",
+                json={
+                    "user": self.user,
+                    "token": self.token,
+                },
+                verify=False,
+            )
+            response.raise_for_status()
+            return BytesIO(response.content).getvalue()
+        except requests.HTTPError as e:
+            if response.status_code == 401:
+                raise PlomAuthenticationException() from None
+            elif response.status_code == 404:
+                raise PlomSeriousException(
+                    "Cannot find image file for {}.".format(num)
+                ) from None
+            elif response.status_code == 406:
+                raise PlomTaskChangedError(
+                    "Task {} has been changed by manager.".format(num)
+                ) from None
+            elif response.status_code == 410:
+                raise PlomTaskDeletedError(
+                    "Task {} has been deleted by manager.".format(num)
+                ) from None
+            else:
+                raise PlomSeriousException(
+                    "Some other sort of error {}".format(e)
+                ) from None
+        finally:
+            self.SRmutex.release()