+++ plom/client/marker.py
@@ -1273,13 +1273,24 @@ class MarkerClient(QWidget):
         if len(self.examModel.getOriginalFiles(task)) > 0:
             return True
 
-        # TODO: plom file is lovely json: why we pack it around as binary bytes?
+        assert task[0] == "q"
+        assert task[5] == "g"
+        num = int(task[1:5])
+        question = int(task[6:])
+        assert question == self.question
+
         try:
-            [page_metadata, anImage, plomfile_data] = self.msgr.MrequestImages(
-                task, self.examModel.getIntegrityCheck(task)
+            integrity = self.examModel.getIntegrityCheck(task)
+            plomdata = self.msgr.get_annotations(
+                num, self.question, edition=None, integrity=integrity
+            )
+            annotated_image = self.msgr.get_annotations_image(
+                num, self.question, edition=plomdata["annotation_edition"]
             )
         except (PlomTaskChangedError, PlomTaskDeletedError) as ex:
             # TODO: better action we can take here?
+            # TODO: the real problem here is that the full_pagedata is potentially out of date!
+            # TODO: we also need (and maybe already have) a mechanism to invalidate existing annotations
             ErrorMessage(
                 '<p>The task "{}" has changed in some way by the manager; it '
                 "may need to be remarked.</p>\n\n"
@@ -1295,21 +1306,16 @@ class MarkerClient(QWidget):
             self.throwSeriousError(e)
             return False
 
-        # TODO: not trivial to replace page_metadata with the full_pagedata:
-        # "included" column means different things.  Maybe we need to
-        # pull down the DB's Annotation records, applied to read-only
-        # image data.  Maybe a shortcut is grab from the plom file.
-        num = int(task[1:5])
+        # Not yet easy to use full_pagedata to build src_img_data (e.g., "included"
+        # column means different things).  Instead, extract from .plom file.
         full_pagedata = self.msgr.MrequestWholePaperMetadata(num, self.question)
         for r in full_pagedata:
             r["local_filename"] = None
         self._full_pagedata[num] = full_pagedata
 
-        paperDir = tempfile.mkdtemp(prefix=task + "_", dir=self.workingDirectory)
-        log.debug("create paperDir {} for already-graded download".format(paperDir))
-
-        src_img_data = [{"id": x[0], "md5": x[1]} for x in page_metadata]
-        del page_metadata
+        log.info("importing source image data (orientations etc) from .plom file")
+        # filenames likely stale
+        src_img_data = plomdata["base_images"]
 
         # Image names = "<task>.<imagenumber>.<extension>"
         # TODO: use server filename from server_path_filename
@@ -1322,33 +1328,19 @@ class MarkerClient(QWidget):
             for r in full_pagedata:
                 if r["md5"] == row["md5"]:
                     r["local_filename"] = tmp
-        # Parse PlomFile early for orientation data
-        plomdata = json.loads(io.BytesIO(plomfile_data).getvalue())
-        db_ids = plomdata.get("database_ids")
-        if db_ids:
-            # if present, must match expected: TODO may delete later.
-            db_id2 = [x["id"] for x in src_img_data]
-            assert db_ids == db_id2, f".plom file IDs={db_ids} does not match {db_id2}"
-        log.info("importing orientations from plom file")
-        for i, row in enumerate(src_img_data):
-            if db_ids:
-                assert (
-                    db_ids[i] == row["id"]
-                ), "sanity check failure: orientation data in wrong order?"
-            row["orientation"] = plomdata["orientations"][i]
 
         self.examModel.setOriginalFilesAndData(task, src_img_data)
 
-        if anImage is None:
-            return True
-
+        paperDir = tempfile.mkdtemp(prefix=task + "_", dir=self.workingDirectory)
+        log.debug("create paperDir {} for already-graded download".format(paperDir))
         self.examModel.setPaperDirByTask(task, paperDir)
         aname = os.path.join(paperDir, "G{}.png".format(task[1:]))
         pname = os.path.join(paperDir, "G{}.plom".format(task[1:]))
         with open(aname, "wb+") as fh:
-            fh.write(anImage)
-        with open(pname, "wb+") as fh:
-            fh.write(plomfile_data)
+            fh.write(annotated_image)
+        with open(pname, "w") as f:
+            json.dump(plomdata, f, indent="  ")
+            f.write("\n")
         self.examModel.setAnnotatedFile(task, aname, pname)
         return True