+++ plom/client/randoMarker.py
@@ -10,7 +10,7 @@ This is a very very cut-down version of Annotator, used to
 automate some random marking of papers.
 """
 
-__copyright__ = "Copyright (C) 2020 Andrew Rechnitzer and others"
+__copyright__ = "Copyright (C) 2020-2021 Andrew Rechnitzer and others"
 __credits__ = "The Plom Project Developers"
 __license__ = "AGPL-3.0-or-later"
 
@@ -18,6 +18,7 @@ import argparse
 import getpass
 import json
 import os
+from pathlib import Path
 import random
 import sys
 import tempfile
@@ -66,19 +67,16 @@ class RW:
 
 class SceneParent(QWidget):
     def __init__(self, question, maxMark):
-        super(SceneParent, self).__init__()
+        super().__init__()
         self.view = PageView(self)
         self.ink = QPen(Qt.red, 2)
         self.question = question
         self.maxMark = maxMark
         self.rubric_widget = RW()  # a dummy class needed for compat with pagescene.
 
-    def doStuff(self, imageNames, saveName, maxMark, markStyle):
+    def doStuff(self, src_img_data, saveName, maxMark, markStyle):
         self.saveName = saveName
-        src_img_data = []
-        for f in imageNames:
-            src_img_data.append({"filename": f, "orientation": 0})
-        self.imageFiles = imageNames
+        self.src_img_data = src_img_data
 
         self.scene = PageScene(self, src_img_data, saveName, maxMark, None)
         self.view.connectScene(self.scene)
@@ -87,7 +85,7 @@ class SceneParent(QWidget):
         lst = self.scene.pickleSceneItems()  # newest items first
         lst.reverse()  # so newest items last
         plomDict = {
-            "fileNames": [os.path.basename(fn) for fn in self.imageFiles],
+            "base_images": self.src_img_data,
             "saveName": os.path.basename(self.saveName),
             "markState": self.scene.getMarkingState(),
             "maxMark": self.maxMark,
@@ -179,9 +177,6 @@ class SceneParent(QWidget):
         self.pickleIt()
         return self.scene.score, self.scene.get_rubrics_from_page()
 
-    def changeMark(self, delta):
-        self.score += delta
-
     def refreshDisplayedMark(self, score):
         # needed for compat with pagescene.py
         pass
@@ -191,24 +186,16 @@ class SceneParent(QWidget):
         pass
 
 
-def annotatePaper(question, maxMark, task, imageList, aname, tags):
+def annotatePaper(question, maxMark, task, src_img_data, aname, tags):
     print("Starting random marking to task {}".format(task))
-    # Image names = "<task>.<imagenumber>.<ext>"
-    with tempfile.TemporaryDirectory() as td:
-        inames = []
-        for i in range(len(imageList)):
-            tmp = os.path.join(td, "{}.{}.image".format(task, i))
-            inames.append(tmp)
-            with open(tmp, "wb+") as fh:
-                fh.write(imageList[i])
-        annot = SceneParent(question, maxMark)
-        annot.doStuff(inames, aname, maxMark, random.choice([2, 3]))
-        annot.doRandomAnnotations()
-        # Issue #1391: settle annotation events, avoid races with QTimers
-        Qapp.processEvents()
-        time.sleep(0.25)
-        Qapp.processEvents()
-        return annot.doneAnnotating()
+    annot = SceneParent(question, maxMark)
+    annot.doStuff(src_img_data, aname, maxMark, random.choice([2, 3]))
+    annot.doRandomAnnotations()
+    # Issue #1391: settle annotation events, avoid races with QTimers
+    Qapp.processEvents()
+    time.sleep(0.25)
+    Qapp.processEvents()
+    return annot.doneAnnotating()
 
 
 def startMarking(question, version):
@@ -222,19 +209,24 @@ def startMarking(question, version):
         # print("Trying to claim next ask = ", task)
         try:
             image_metadata, tags, integrity_check = messenger.MclaimThisTask(task)
-        except PlomTakenException as e:
+        except PlomTakenException:
             print("Another user got task {}. Trying again...".format(task))
             continue
 
-        image_md5s = [row[1] for row in image_metadata]
-        imageList = []
-        for row in image_metadata:
-            imageList.append(messenger.MrequestOneImage(task, row[0], row[1]))
+        src_img_data = [
+            {"id": r[0], "md5": r[1], "orientation": 0} for r in image_metadata
+        ]
         with tempfile.TemporaryDirectory() as td:
+            for i, r in enumerate(src_img_data):
+                obj = messenger.MrequestOneImage(task, r["id"], r["md5"])
+                tmp = os.path.join(td, f"{task}.{i}.image")
+                with open(tmp, "wb") as f:
+                    f.write(obj)
+                r["filename"] = tmp
             aFile = os.path.join(td, "argh.png")
             plomFile = aFile[:-3] + "plom"
             score, rubrics = annotatePaper(
-                question, maxMark, task, imageList, aFile, tags
+                question, maxMark, task, src_img_data, aFile, tags
             )
             print("Score of {} out of {}".format(score, maxMark))
             messenger.MreturnMarkedTask(
@@ -248,7 +240,7 @@ def startMarking(question, version):
                 plomFile,
                 rubrics,
                 integrity_check,
-                image_md5s,
+                [r["md5"] for r in src_img_data],
             )