+++ plom/db/db_identify.py
@@ -9,7 +9,8 @@ import peewee as pw
 
 from plom.rules import censorStudentNumber as censorID
 from plom.rules import censorStudentName as censorName
-from plom.db.tables import *
+from plom.db.tables import plomdb
+from plom.db.tables import DNMGroup, DNMPage, Group, IDGroup, IDPage, Test, User
 
 
 log = logging.getLogger("DB")
@@ -33,7 +34,7 @@ def IDcountAll(self):
             )
             .count()
         )
-    except Group.DoesNotExist:
+    except pw.DoesNotExist:
         return 0
 
 
@@ -49,7 +50,7 @@ def IDcountIdentified(self):
             )
             .count()
         )
-    except IDGroup.DoesNotExist:
+    except pw.DoesNotExist:
         return 0
 
 
@@ -67,7 +68,7 @@ def IDgetNextTask(self):
                 .get()
             )
             # note - test need not be all scanned, just the ID pages.
-        except IDGroup.DoesNotExist:
+        except pw.DoesNotExist:
             log.info("Nothing left on ID to-do pile")
             return None
 
@@ -127,8 +128,8 @@ def IDgetDoneTasks(self, user_name):
     return idList
 
 
-def IDgetImage(self, user_name, test_number):
-    """Return ID page images (+ Lpages) of test #test_number to user."""
+def IDgetImages(self, user_name, test_number):
+    """Return ID page images of a paper."""
     uref = User.get(name=user_name)
     # since user authenticated, this will always return legit ref.
 
@@ -147,6 +148,21 @@ def IDgetImage(self, user_name, test_number):
     for p in iref.idpages.order_by(IDPage.order):
         rval.append(p.image.file_name)
     log.debug("Sending IDpages of test {} to user {}".format(test_number, user_name))
+    return rval
+
+
+def ID_get_donotmark_images(self, test_number):
+    """Return the DoNotMark page images of a paper."""
+    tref = Test.get_or_none(Test.test_number == test_number)
+    if tref is None:
+        return [False, "NoTest"]
+    iref = tref.dnmgroups[0]
+    if iref.group.scanned is False:
+        return [False, "NoScan"]
+    rval = [True]
+    for p in iref.dnmpages.order_by(DNMPage.order):
+        rval.append(p.image.file_name)
+    log.debug(f"Sending DNMpages of test {test_number}")
     return rval