+++ plom/server/plomServer/serverID.py
@@ -55,18 +55,29 @@ def IDgetDoneTasks(self, username):
     return self.DB.IDgetDoneTasks(username)
 
 
-def IDgetImage(self, username, test_number):
-    """Respond with a list of image paths for an already ID'd paper.
+def IDgetImages(self, username, test_number):
+    """Respond with a list of image paths the ID pages of a paper.
 
     Args:
         username (str): Username requesting ID'd paper's image.
-        test_number (str): Test ID number.
+        test_number (str): Test number.
 
     Returns:
-        list: True/False plus a list of the image paths for ID'd task.
+        list: True/False plus a list of the image paths or a short error code.
     """
+    return self.DB.IDgetImages(username, test_number)
 
-    return self.DB.IDgetImage(username, test_number)
+
+def ID_get_donotmark_images(self, paper_number):
+    """Respond with a list of image paths for the Do Not Mark pages of a paper.
+
+    Args:
+        test_number (str): Test number.
+
+    Returns:
+        list: True/False plus a list of the image paths or a short error code.
+    """
+    return self.DB.ID_get_donotmark_images(paper_number)
 
 
 def IDclaimThisTask(self, username, test_number):